Correcting data types

+++ scripts/sample/SALibRepastParams.py
@@ -8,8 +8,8 @@ from xml.etree import ElementTree as et
 params = {
 			"epsilon":{				"type":"constant", "data_type":"double", 	"value":"2.5", "bounds":[0,4], "dist":"unif"},
 			"lambda":{				"type":"constant", "data_type":"double", 	"value":"0.8", "bounds":[0,1], "dist":"unif"},
-			"addPedTicks":{			"type":"constant", "data_type":"double", 	"value":"50", "bounds":[10,100], "dist":"unif"},
-			"addVehicleTicks":{		"type":"constant", "data_type":"double", 	"value":"400", "bounds":[200,600], "dist":"unif"},
+			"addPedTicks":{			"type":"constant", "data_type":"int", 		"value":"50", "bounds":[10,100], "dist":"unif"},
+			"addVehicleTicks":{		"type":"constant", "data_type":"int", 		"value":"400", "bounds":[200,600], "dist":"unif"},
 			"gamma":{				"type":"constant", "data_type":"double", 	"value":"0.9", "bounds":[0,1], "dist":"unif"},
 			"randomSeed":{			"type":"constant", "data_type":"int", 		"value":"1", "bounds":[1,100], "dist":"unif"},
 			"alpha":{				"type":"constant", "data_type":"double", 	"value":"0.5", "bounds":[0,1], "dist":"unif"},
